requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
psycopg2
python-dotenv
strawberry-graphql[fastapi]
asyncpg

 